change(form-field): Move the component from the account folder to the common folder and rename the file
